src: lib: Move functionality to compress_hw.c
Add compress_hw.c to handle all operations associated
with HW compress nodes. Move handling to compress_hw
using compress_hw_ops.
