+++ src/lib/Makefile.am
@@ -1,5 +1,5 @@
 tinycompressdir = $(libdir)
 
 tinycompress_LTLIBRARIES = libtinycompress.la
-libtinycompress_la_SOURCES = compress.c
+libtinycompress_la_SOURCES = compress.c compress_hw.c
 libtinycompress_la_CFLAGS = -I$(top_srcdir)/include
